Random player capped at a max of 3 ants at the same time
I've modified the Random AI player so that it won't have more than 3 ants
on the field at a time.  This is in response to the fact that workers
can not attack anymore.  As a result, the previous random players was
building an army of worker ants when played against itself.  This resulted
in overlong random vs. random games.