Added getNextState method and fixed a minor bug.
The getNextState method will be used for search algorithms.

The bug was fixed by Jarret Oney who writes:  I think I found a slight error in the Utils file, specifically in the createPathTowards function. On finding the next coordinate in the path, the distance to the target is never updated for that new coord. This can cause a drone (and only the drone) to make an inefficient path as the algorithm it uses is greedy. For example: a drone with an initial distance of 7 finds a spot to bring the distance to 6, then a spot to bring it to 5, but then another spot to bring it to 6 which is still valid since it is checking against 7 the whole time.